src/main/java/com/task/routeadvisor/service/CountryService.java
package com.task.routeadvisor.service;

import com.task.routeadvisor.entity.Country;
import com.task.routeadvisor.repository.CountryRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import org.springframework.web.server.ResponseStatusException;

@Service
@RequiredArgsConstructor
public class CountryService {

    private final CountryRepository countryRepository;

    @Transactional(readOnly = true)
    public Country findCountryByCode(final String code) {
        return countryRepository.findByCode(code)
            .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Country '" + code + "' not found"));
    }

}
